+++ src/app/touch_manager.cpp
@@ -20,6 +20,8 @@
 #include "drivers/axs15231b_touch_driver.h"
 #elif TOUCH_DRIVER == TOUCH_DRIVER_CST816S_ESP_PANEL
 #include "drivers/esp_panel_cst816s_touch_driver.h"
+#elif TOUCH_DRIVER == TOUCH_DRIVER_GT911
+#include "drivers/gt911_touch_driver.h"
 #endif
 
 // Global instance
@@ -84,6 +86,8 @@ void TouchManager::init() {
 		driver = new AXS15231B_TouchDriver();
 		#elif TOUCH_DRIVER == TOUCH_DRIVER_CST816S_ESP_PANEL
 		driver = new ESPPanel_CST816S_TouchDriver();
+		#elif TOUCH_DRIVER == TOUCH_DRIVER_GT911
+		driver = new GT911_TouchDriver();
 		#else
 		#error "No touch driver selected or unknown driver type"
 		#endif